feat(resultpage): replace result card. configurate component. add route
