+++ src/config/routeConfig/routeConfig.tsx
@@ -3,11 +3,13 @@ import { NotFoundPage } from "../../pages/NotFoundPage";
 import { TestPage } from "../../pages/TestPage";
 import { MainPage } from "../../pages/MainPage";
 import { SearchPage } from "../../pages/SearchPage";
+import { ResultPage } from "../../pages/ResultPage";
 import { AppRoutes } from "../types";
 
 export const RoutePath: Record<AppRoutes, string> = {
   [AppRoutes.MAIN_PAGE]: "/",
   [AppRoutes.SEARCH_PAGE]: "/search",
+  [AppRoutes.RESULT_PAGE]: "/results",
   [AppRoutes.TEST_PAGE]: "/test",
   [AppRoutes.NOT_FOUND]: "*",
 };
@@ -20,6 +22,10 @@ export const routeConfig: Record<AppRoutes, RouteProps> = {
   [AppRoutes.SEARCH_PAGE]: {
     path: RoutePath.search_page,
     element: <SearchPage initialStep={0} />,
+  },
+  [AppRoutes.RESULT_PAGE]: {
+    path: RoutePath.result_page,
+    element: <ResultPage />,
   },
   [AppRoutes.TEST_PAGE]: {
     path: RoutePath.test_page,
