+++ src/config/types/index.ts
@@ -1,6 +1,7 @@
 export enum AppRoutes {
   MAIN_PAGE = "main_page",
   SEARCH_PAGE = "search_page",
+  RESULT_PAGE = "result_page",
   TEST_PAGE = "test_page",
   NOT_FOUND = "not_found",
 }
