+++ src/pages/ResultPage/index.tsx
@@ -1,0 +1,1 @@
+export { ResultPageAsync as ResultPage } from "./ui/ResultPage.async.tsx";
