+++ src/pages/ResultPage/ui/ResultCard/ResultCard.module.scss
@@ -1,4 +1,4 @@
-@import "../../app/styles/mixins";
+@import "../../../../app/styles/mixins";
 
 .section {
   box-sizing: border-box;
