+++ src/pages/ResultPage/ui/ResultHeader/ResultHeader.module.scss
@@ -1,0 +1,7 @@
+.header {
+  padding: 32px 32px 0;
+  display: flex;
+  align-items: center;
+  width: 100%;
+  justify-content: space-between;
+}
