+++ src/pages/ResultPage/ui/ResultPage.async.tsx
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const ResultPageAsync = lazy(async () => import("./ResultPage.tsx"));
